marketing-bot/requirements.txt
facebook-business>=17.0.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
apify-client>=2.0.0
aiohttp>=3.9.0
gspread>=6.0.0
google-auth>=2.0.0
